Installed pinia to store weatherdata
