+++ package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "axios": "^1.7.9",
+    "pinia": "^3.0.0",
     "vue": "^3.5.13",
     "vue-router": "^4.5.0"
   },
